AP-21337: use GoogleCredential through AccessTokenAccessor interface in Google nodes
The Google Authenticator nodes (old and new) still produce a GoogleCredential. However
the downstream nodes need to be compatible with the credentials produced by the
Secret Retriever node as well. GoogleCredential is still required as its own type
because Cloud Storage signed URLs need access to the underlying private key
of a ServiceAccountCredential.

CredentialUtil provides helper methods to adapt an AccessTokenAccessor to a Google
OAuth2Accessor, unwrapping an underlying ServiceAccountCredential if possible.

AP-21337 (Classloading hack for BigQuery connector)